+++ docs/source/whatsnew.rst
@@ -4,6 +4,9 @@ What's New
 v2.1.0
 ------
 - The :class:`wxcadm.XSICallQueue` class was added for control of Call Queue calls
+- New method :meth:`wxcadm.Webex.get_person_by_email()` which does what the Org-level method does, but searches across all Orgs that the user can manage.
+- Failed "pushes" to Webex for user data no longer raise an exception. They now return False to prevent blocking in scripts.
+- :meth:`wxcadm.Person.push_vm_config()` now supports a vm_config dict rather than modifying the :attr:`Person.vm_config` attribute directly.
 
 v2.0.0
 ------